+++ src/datasources/results/components/editors/time-range/TimeRangeControls.tsx
@@ -26,7 +26,7 @@ export function TimeRangeControls({query, handleQueryChange}: Props) {
 
   return (
     <div className="horizontal-control-group">
-      <InlineField label="Use time range" tooltip={tooltips.useTimeRange} labelWidth={25}>
+      <InlineField label="Use time range" tooltip={tooltips.useTimeRange} labelWidth={26}>
         <InlineSwitch
           onChange={event => onUseTimeRangeChecked(event.currentTarget.checked)}
           value={query.useTimeRange}